included orders, fixed ddl

+++ A3/stockExchange_DDL.sql
@@ -213,10 +213,20 @@ alter table handlesCommodities add constraint hc_fk_cid foreign key(c_id) refere
 
 alter table phoneNumber add constraint pn_fk_brokerid foreign key(broker_id) references brokers(broker_id);
 
+create function updateStockValue(stockid int, stockPrice DECIMAL(10,2))
+returns void
+language plpgsql
+as
+$$
+declare
+   return_value int;
+begin
+    UPDATE stocks SET currentValue = stockPrice where s_id = stockid;
+end;
+$$;
 
 
-
-create function updateBuyerHoldings(buyerid int, stockid int, stockquantity int)
+create function updateBuyerHoldings(buyerid int, stockid int, stockquantity int, StockPrice DECIMAL(10,2))
 returns void
 language plpgsql
 as
@@ -234,13 +244,32 @@ begin
         INSERT INTO holdsStocks values(buyerid, stockid, stockquantity);
         return_value := 1;
     END IF;
-   
---    return return_value;
+    UPDATE investorsAndTraders SET marginAvailable = marginAvailable - StockPrice * stockquantity where i_id = buyerid;
 end;
 $$;
 
+create function updateBuyerHoldingsInQueue(buyerid int, stockid int, stockquantity int, StockPrice DECIMAL(10,2))
+returns void
+language plpgsql
+as
+$$
+declare
+   return_value int;
+   stockheld char(100);
+begin
+   return_value := 0;
+    SELECT exists into stockheld(SELECT * FROM holdsStocks WHERE s_id=stockid and i_id = buyerid);
+    if(stockheld = 'true') THEN
+        UPDATE holdsStocks SET quantity = quantity + stockquantity WHERE s_id = stockid and i_id = buyerid;
+        return_value := 1;
+    ELSE
+        INSERT INTO holdsStocks values(buyerid, stockid, stockquantity);
+        return_value := 1;
+    END IF;
+end;
+$$;
 
-create function updateSellerHoldings(Sellerid int, stockid int, stockquantity int)
+create function updateSellerHoldings(Sellerid int, stockid int, stockquantity int, StockPrice DECIMAL(10,2))
 returns void
 language plpgsql
 as
@@ -255,6 +284,8 @@ begin
     ELSE
         UPDATE holdsStocks SET quantity = quantity - stockquantity WHERE s_id = stockid and i_id = sellerid;
     END IF;
+    --update margin
+    UPDATE investorsAndTraders SET marginAvailable = marginAvailable + StockPrice * stockquantity where i_id = Sellerid;
 end;
 $$;
 
@@ -274,79 +305,44 @@ DECLARE return_value int;
     buyer_margin DECIMAL(10,2);
     stockheld char(10);
     sellerid int;
+    quantityHeld int;
 BEGIN 
 
     SELECT currentValue into stock_value FROM stocks WHERE s_id = stockid;
     SELECT marginAvailable into buyer_margin FROM investorsAndTraders WHERE i_id = buyerid;
+    SELECT quantity into quantityHeld FROM holdsStocks WHERE i_id = buyerid and s_id = stockid;
 
     IF((stock_value*buyquantity) < buyer_margin)  THEN
         
         Select exists into seller_present(SELECT seller_id, stock_id, sell_price, sell_quantity from sell_table WHERE stock_id=stockid);
         SELECT sell_quantity into sellquantitycompute from sell_table WHERE stock_id=stockid and sell_price=buyprice;
         SELECT seller_id into sellerid from sell_table WHERE stock_id=stockid and sell_price=buyprice;
+        SELECT sell_quantity into sellquantitycompute from sell_table WHERE stock_id=stockid and sell_price=buyprice;
         
-        if(seller_present = 'true' and sellerid != buyerid) THEN
+        if(seller_present = 'true' and sellerid != buyerid) THEN                        
             
-            SELECT sell_quantity into sellquantitycompute from sell_table WHERE stock_id=stockid and sell_price=buyprice;
-            
-            if(sellquantitycompute-buyquantity = 0) THEN
+            if(sellquantitycompute = buyquantity ) THEN
                 DELETE FROM sell_table where stock_id=stockid and sell_price=buyprice;
-                SELECT exists into stockheld(SELECT * FROM holdsStocks WHERE s_id=stockid and i_id = buyerid);
-
-                --handle deletion of quantity in seller
-                DELETE FROM holdsStocks where s_id = stockid and i_id = sellerid;
-                
-
-                PERFORM updateSellerHoldings(sellerid, stockid, sellquantitycompute);
-
-                --update margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable - buyprice * buyquantity where i_id = buyerid;
-
-                if(stockheld = 'true') THEN
-                    UPDATE holdsStocks SET quantity = quantity + buyquantity WHERE s_id = stockid and i_id = buyerid;
-                ELSE
-                    INSERT INTO holdsStocks values(buyerid, stockid, buyquantity);
-                END IF;
+           
+                PERFORM updateSellerHoldings(sellerid, stockid, sellquantitycompute, buyprice);
+                PERFORM updateBuyerHoldings(buyerid, stockid, buyquantity, buyprice);
             
             ELSIF ( sellquantitycompute > buyquantity ) THEN
+                UPDATE sell_table SET sell_quantity = sell_quantity - buyquantity where stock_id = stockid and seller_id= sellerid;                
 
-                raise notice 'Value: %', sellerid;
-
-                UPDATE sell_table SET sell_quantity = sell_quantity - buyquantity where stock_id = stockid and seller_id= sellerid;
-                
-                SELECT exists into stockheld(SELECT * FROM holdsStocks WHERE s_id=stockid and i_id = buyerid);
-                raise notice 'buy quantity: %', buyquantity;
-                PERFORM updateSellerHoldings(sellerid, stockid, buyquantity);
-
-                --update margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable - buyprice * buyquantity where i_id = buyerid;
-
-                -- for the buyer
-                if(stockheld = 'true') THEN
-                    UPDATE holdsStocks SET quantity = quantity + buyquantity WHERE s_id = stockid and i_id = buyerid;
-                ELSE                    
-                    INSERT INTO holdsStocks values(buyerid, stockid, buyquantity);
-                END IF;
-                
+                PERFORM updateSellerHoldings(sellerid, stockid, buyquantity, buyprice);
+                PERFORM updateBuyerHoldings(buyerid, stockid, buyquantity, buyprice);
             
             ELSE -- sellquantitycompute < buyquantity
-                PERFORM updateSellerHoldings(sellerid, stockid, sellquantitycompute);
                 DELETE FROM sell_table where stock_id=stockid and sell_price=buyprice and seller_id=sellerid;
                 INSERT INTO buy_table(buyer_id, stock_id, buy_price, buy_quantity) values (buyerid, stockid, buyprice, buyquantity-sellquantitycompute);
-                
-                SELECT exists into stockheld(SELECT * FROM holdsStocks WHERE s_id=stockid and i_id = buyerid);
 
-                --update margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable - buyprice * sellquantitycompute where i_id = buyerid;
+                PERFORM updateSellerHoldings(sellerid, stockid, sellquantitycompute, buyprice);
+                PERFORM updateBuyerHoldings(buyerid, stockid, sellquantitycompute, buyprice);
+            END IF;
 
-                if(stockheld = 'true') THEN
-                    UPDATE holdsStocks SET quantity = quantity + sellquantitycompute WHERE s_id = stockid and i_id = buyerid;
-                ELSE
-                    INSERT INTO holdsStocks values(buyerid, stockid, sellquantitycompute);
-                END IF;
-            END IF;          
-        
-        return_value := 1;
+            PERFORM updateStockValue(stockid, buyprice);
+            return_value := 1;
         
         ELSE
             INSERT INTO buy_table(buyer_id, stock_id, buy_price, buy_quantity) values (buyerid, stockid, buyprice, buyquantity); 
@@ -392,101 +388,43 @@ begin
 
 
     if(stockheld = 'true' and sellquantity <= numberofstocksheld) THEN
-
         
         SELECT exists into buyer_present(SELECT * from buy_table where stock_id = stockid and buy_price = sellprice);
         SELECT buy_quantity into buyquantitycompute from buy_table where stock_id = stockid and buy_price = sellprice;
         SELECT buyer_id into buyerid from buy_table where stock_id = stockid and buy_price = sellprice;
 
-        if(buyer_present = 'true') THEN
-
+        if(buyer_present = 'true' and sellerid != buyerid) THEN
 
             if(buyquantitycompute = sellquantity ) THEN
                 DELETE from buy_table where stock_id = stockid and buyer_id = buyerid;
-
                 -- handle waiting things too
-                PERFORM updateBuyerHoldings(buyerid, stockid, buyquantitycompute);
-
-
-                -- DELETE FROM buy_table where s_id = stockid and buy_price = sellprice;
-                --updating margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable + sellprice * sellquantity where i_id = sellerid;
-                
-                if(numberofstocksheld = sellquantity) THEN
-                    DELETE FROM holdsStocks where s_id = stockid and i_id = sellerid;
-                    
-
-                ELSIF (numberofstocksheld > sellquantity) THEN
-                    UPDATE holdsStocks SET quantity = quantity - sellquantity where s_id = stockid and i_id = sellerid;
-                    
-                ELSE
-                    raise notice 'You cannot sell more than what you own';
-                    
-                END IF;
-                
+                PERFORM updateBuyerHoldingsInQueue(buyerid, stockid, buyquantitycompute, sellprice);                
+                PERFORM updateSellerHoldings(sellerid, stockid, buyquantitycompute, sellprice);                
 
             ELSIF (buyquantitycompute > sellquantity ) THEN
                 UPDATE buy_table set buy_quantity = buy_quantity - sellquantity where stock_id = stockid and buy_price = sellprice;
-                PERFORM updateBuyerHoldings(buyerid, stockid, sellquantity); -- as buyquantity is more
+                PERFORM updateBuyerHoldingsInQueue(buyerid, stockid, sellquantity, sellprice);
+                PERFORM updateSellerHoldings(sellerid, stockid, sellquantity, sellprice); 
                 
-                
-                --updating margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable + sellprice * sellquantity where i_id = sellerid;
 
-
-                if(numberofstocksheld = sellquantity) THEN
-                    DELETE FROM holdsStocks where s_id = stockid and i_id = sellerid;
-                    
-
-                ELSIF (numberofstocksheld > sellquantity) THEN
-                    UPDATE holdsStocks SET quantity = quantity - sellquantity where s_id = stockid and i_id = sellerid;
-                    
-                ELSE
-                    raise notice 'You cannot sell more than what you own';
-                    
-                END IF;
-
-
-            ELSE
+            ELSE --buyquantitycompute < sellquantity
                 DELETE from buy_table where stock_id = stockid and buyer_id = buyerid;
-                PERFORM updateBuyerHoldings(buyerid, stockid, buyquantitycompute);
+                PERFORM updateBuyerHoldingsInQueue(buyerid, stockid, buyquantitycompute, sellprice);
+                PERFORM updateSellerHoldings(sellerid, stockid, buyquantitycompute, sellprice); 
 
-
-                --updating margin
-                UPDATE investorsAndTraders SET marginAvailable = marginAvailable + sellprice * buyquantitycompute where i_id = sellerid;
-
-                if(numberofstocksheld = sellquantity) THEN
-                    DELETE FROM holdsStocks where s_id = stockid and i_id = sellerid;
-                    
-
-                ELSIF (numberofstocksheld > sellquantity) THEN
-                    UPDATE holdsStocks SET quantity = quantity - buyquantitycompute where s_id = stockid and i_id = sellerid;
-                    
-                ELSE
-                    raise notice 'You cannot sell more than what you own';
-                    
-                END IF;
-
-                INSERT into sell_table values (sellerid, stockid, sellprice, sellquantity - buyquantitycompute );
-                
-
-
+                INSERT into sell_table values (sellerid, stockid, sellprice, sellquantity - buyquantitycompute );               
             END IF;
             
+            PERFORM updateStockValue(stockid, sellprice);
             return_value := 1;
 
-
-
-
         ELSE -- no buyers
-            INSERT into sell_table values (sellerid, stockid, sellprice, sellquantity );
-            --update margin
-            UPDATE investorsAndTraders SET marginAvailable = marginAvailable + sellprice * sellquantity where i_id = sellerid;
+            INSERT into sell_table values (sellerid, stockid, sellprice, sellquantity );            
             return_value := 2;
         END IF;
     
     ELSE
-        raise notice 'You dont have that stock';
+        raise notice 'You dont have that stock or trying to sell more than what you own';
         return_value := 0;
 
     END IF;
